requirements.txt
google-generativeai
langchain
langchain-community 
faiss-cpu
chromadb
tiktoken
pandas
requests
python-dotenv
tabulate
sentence-transformers
